Fix incorrect Bundle configuration flag
On my machine, Bundle is unable to build this project as a user (without
root privileges) because the configuration flag to have Bundle run in a
user directory is not set correctly. In particular, the project
currently has `bundle config --local path set .vendor/bundle`, but it
should be `bundle config --local path .vendor/bundle` (without the
`set`). In this commit, I have removed the extraneous `set`, and the
project correctly serves on my machine.